tools(test): switch prealign test to vedo.Mesh I/O to match new API
- Generate meshes with vedo (Sphere), pass vedo.Mesh directly to _prealign_interactive_np.
- Compute before/after metrics from mesh.points().
- Save adjusted source via vedo .write().
- Post-check visualization moved to vedo (target gray, adjusted blue, original red wireframe).
